Conflicts:
	src/gui/Menu2D.h

+++ src/gui/Model.h
@@ -21,7 +21,7 @@ public:
 	// interface methods
 	void loadScene();
 	void draw();
-	void setCorrectionValues(int localX, int localY, int localZ, float scaleFactor, int rotateX, int rotateY, int rotateZ);
+	void setCorrectionValues(float localX, float localY, float localZ, float scaleFactor, float rotateX, float rotateY, float rotateZ);
 	void rotateY(int degree);
 	void setColor(Color color);
 	void setPosition(int globalX, int globalY, int globalZ);
